Adds a working Android Pay button
This commit enables Android Pay in the app. The app enables the 'Buy with Android Pay' button if it can, and demonstrates a typical payment flow, showing a PaymentMethodToken at the end.

Adds a working Android Pay button

This commit enables Android Pay in the app. The app enables the 'Buy with Android Pay' button if it can, and demonstrates a typical payment flow, showing a PaymentMethodToken at the end.